Move: cleanup memory cache
