+++ src/heretic/main.py
@@ -48,10 +48,9 @@ from .analyzer import Analyzer
 from .config import QuantizationMethod, Settings
 from .evaluator import Evaluator
 from .model import AbliterationParameters, Model, get_model_class
-from .system import get_accelerator_info
+from .system import empty_cache, get_accelerator_info
 from .utils import (
     create_reproduce_folder,
-    empty_cache,
     format_duration,
     get_readme_intro,
     get_trial_parameters,
